[#154111303] Architect wants to force unique integer for chord position in phase or audio, in order to segment any phase or audio into discreet blocks for analysis.   * [core/migrations] V53 modify tables `phase_chord` `link_chord` `audio_chord` MODIFY COLUMN `position` INT NOT NOULL   * [core/model] PhaseChord, LinkChord, and AudioChord models get Integer makeover

+++ core/src/main/java/io/xj/core/work/basis/impl/BasisImpl.java
@@ -383,10 +383,10 @@ public class BasisImpl implements Basis {
   }
 
   @Override
-  public Chord chordAt(double position) throws Exception {
+  public Chord chordAt(int position) throws Exception {
     // default to returning a chord based on the link key, if nothing else is found
     Chord foundChord = Chord.of(link().getKey());
-    Double foundPosition = null;
+    Integer foundPosition = null;
 
     // we assume that these chords are in order of position ascending (see: LinkChordDAO.readAllExpectedWork)
     for (LinkChord linkChord : linkChords()) {
